src/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.0
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0
safetensors==0.3.3
torch==2.0.1
pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.2
httpx==0.24.1
